### 2024-06-06 #6846 @Kevin-Mattheus-Moerman
@petrableeker thanks for this submission. I am the AEiC on this track and here to help with initial steps. I am currently reviewing if this work is in scope for JOSS. In particular given the relatively small size of the project I've just flagged this for scope review by our board, to see this conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This scope review should take about 2 weeks to complete.
