+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!--next-version-placeholder-->
 
+## v0.1.2 (23-07-2024)
+
+- Caps in package name at import utils caused an error when using OmicsAnalysis and FeatureSelection.
+
 ## v0.1.1 (04-06-2024)
 
 - The python version caused an error during installation. The python version in pyproject.toml is now changed to 3.9.
