### 2024-07-31 #6846 @jromanowska
Hi, @petrableeker , I've just read the paper and have one comment: I find that the headers for "Summary" and "Statement of need" don't really match the content. I'd suggest headers like "Statement of need" and "Implementation" instead.

Also - do you have any suggestions for (or oppositions) reviewers? Above, our robot has suggested some names, so I'll probably choose some of them, but just wanted to know if you have any comments on this.
