### 2024-09-18 #6846 @jromanowska
> Hi @jromanowska, So far, we could not yet find anyone. One of the authors suggested that Pablo Lopez-Tarifa (lopeztarifa on GitHub) might be able to help us. In the meantime, I'll continue searching.

@LissyDenkers - do you have maybe an email address to Pablo Lopez-Tarifa? Perhaps GitHub notifications did not reach him.
