### 2024-09-20 #6846 @pierrepeterlongo
Hi,
Yes, I accept this review. However, I won't have time before 2 weeks. Is that acceptable?
