### 2024-09-23 2af806e
changed missing and invalid DOIs
