+++ documentation/paper/paper.bib
@@ -162,7 +162,6 @@
 }
 @article{Pedregosa2011,
    author = {Fabian Pedregosa and Gaël Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and Édouard Duchesnay},
-   doi = {10.5555/1953048.2078195},
    journal = {Journal of Machine Learning Research},
    keywords = {Python,model selection,supervised learning,unsupervised learning},
    pages = {2825-2830},
@@ -187,6 +186,7 @@
 
 @article{Seefeldt1995,
    author = {Steven S Seefeldt and Jens Erik Jensen and E. Patrick Fuerst},
+   doi = {10.1017/s0890037x00023253},
    issue = {2},
    journal = {Weed Technology},
    pages = {218-227},
