### 2024-09-27 #7264 @lopeztarifa
Hi @jromanowska,
The software addresses the problem of phenotype linking with omic data. The paper is clear, the repo is full of good examples and the jupyter manuals are easy to follow.
Two remarks,
- The code lacks proper testing.
- Classes are commented on, what about documenting them?

and one suggestion, would it be possible to upgrade the Python version? In a year v3.9 will be obsolete.
