### 2024-09-30 a975f99
update read the docs configuration
