+++ .readthedocs.yml
@@ -3,7 +3,12 @@ version: 2
 
 # Image to use
 build:
-  image: testing
+  os: ubuntu-22.04
+  tools:
+    python: 3.9
+
+sphinx:
+  configuration: docs/conf.py
 
 # Configuration
 python:
