+++ .readthedocs.yml
@@ -5,14 +5,14 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: 3.9
+    python: "3.9"
 
 sphinx:
   configuration: docs/conf.py
 
 # Configuration
 python:
-  version: 3.9
+  version: "3.9"
   install:
     - requirements: docs/requirements.txt
     - method: pip
