### 2024-09-30 6e7ba6a
fix read the docs build error (python version key)
