+++ .readthedocs.yml
@@ -11,9 +11,9 @@ sphinx:
   configuration: docs/conf.py
 
 # Configuration
-python:
-  version: "3.9"
-  install:
-    - requirements: docs/requirements.txt
-    - method: pip
-      path: .
+#python:
+#  version: "3.9"
+#  install:
+#    - requirements: docs/requirements.txt
+#    - method: pip
+#      path: .
