### 2024-09-30 14f3de7
fix read the docs build error (build configuration for Poetry)
