+++ .readthedocs.yml
@@ -6,6 +6,18 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.9"
+  # A special build step for Poetry is needed
+  # instructions from https://test-builds.readthedocs.io/en/poetry/ :
+  jobs:
+    post_create_environment:
+      # Install poetry
+      # https://python-poetry.org/docs/#installing-manually
+      - pip install poetry
+    post_install:
+      # Install dependencies with 'docs' dependency group
+      # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
+      # VIRTUAL_ENV needs to be set manually for now, see #11150
+      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install
 
 # Sphinx configuration
 sphinx:
