### 2024-09-30 042fef7
fix read the docs build error (build configuration for Poetry)
