### 2024-09-30 44e9198
fix read the docs build error (build configuration for Poetry)
