### 2024-09-30 0926e83
fix read the docs build error (build configuration for Poetry)
