### 2024-09-30 2744208
fix read the docs build error (python version used in Poetry virtual env)
