### 2024-09-30 fe292a7
fix read the docs build error (python version used in Poetry virtual env)
