### 2024-09-30 212200b
fix read the docs build error (python version used in Poetry virtual env)
