### 2024-09-30 1eb7aee
fix read the docs build error (python version used in Poetry virtual env)
