+++ .readthedocs.yml
@@ -5,7 +5,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.9"
+    python: "3.9.0"
   # A special build step for Poetry is needed
   # instructions from https://test-builds.readthedocs.io/en/poetry/ :
   jobs:
