### 2024-09-30 1d80200
fix read the docs build error (python version used in Poetry virtual env)
