+++ pyproject.toml
@@ -11,7 +11,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "3.9"
+python = "~3.9"
 seaborn = ">=0.12.1"
 pandas = ">=1.5.1"
 numpy = "1.25.0"
