### 2024-09-30 f969b1a
fix read the docs build error (removed --with docs)
