+++ .readthedocs.yml
@@ -18,7 +18,7 @@ build:
       # https://python-poetry.org/docs/managing-dependencies/#dependency-groups
       # VIRTUAL_ENV needs to be set manually for now, see #11150
       - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry env use 3.9
-      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install --with docs
+      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install
 
 # Sphinx configuration
 sphinx:
