### 2024-09-30 c930e3a
fix read the docs build error (update poetry.lock)
