+++ docs/conf.py
@@ -19,8 +19,7 @@ extensions = [
     "myst_nb",
     "autoapi.extension",
     "sphinx.ext.napoleon",
-    "sphinx.ext.viewcode",
-    "sphinx_autorun"
+    "sphinx.ext.viewcode"
 ]
 autoapi_dirs = ["../src"]
 
