### 2024-09-30 989c05d
fix read the docs build error (change author value to string)
