### 2024-10-01 7ce20f1
working on Read the Docs structure
