+++ docs/index.rst
@@ -1,0 +1,11 @@
+PhenoFeatureFinder
+==================
+
+A python package for linking developmental phenotypes to omics features
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Contents:
+
+    intro
+    docstrings
