+++ docs/docstrings.rst
@@ -1,0 +1,19 @@
+PhenotypeAnalysis
+=================
+
+.. currentmodule:: phenotype_analysis
+.. autofunction:: PhenotypeAnalysis
+
+
+OmicsAnalysis
+=============
+
+.. currentmodule:: omics_analysis
+.. autofunction:: OmicsAnalysis
+
+
+FeatureSelection
+================
+
+.. currentmodule:: feature_selection_using_ml
+.. autofunction:: FeatureSelection
