+++ docs/intro.rst
@@ -1,0 +1,2 @@
+PhenoFeatureFinder
+==================
