+++ docs/docstrings.rst
@@ -1,19 +1,26 @@
+*************
+Documentation
+*************
+
+
 PhenotypeAnalysis
-=================
+#################
 
 .. currentmodule:: phenotype_analysis
 .. autofunction:: PhenotypeAnalysis
 
 
+
 OmicsAnalysis
-=============
+#############
 
 .. currentmodule:: omics_analysis
 .. autofunction:: OmicsAnalysis
 
 
+
 FeatureSelection
-================
+################
 
 .. currentmodule:: feature_selection_using_ml
 .. autofunction:: FeatureSelection
