+++ docs/index.rst
@@ -9,8 +9,6 @@ Contents
 
     intro.rst
     docstrings.rst
-    ../documentation/examples/
-    ../documentation/manuals/
     changelog.md
     contributing.md
     conduct.md
