+++ docs/index.md
@@ -1,11 +1,12 @@
+```{include} intro.rst
+```
+
 ```{toctree}
 :maxdepth: 2
 :hidden:
 
 intro.rst
 docstrings.rst
-../documentation/examples/
-../documentation/manuals/
 changelog.md
 contributing.md
 conduct.md
