+++ src/PhenoFeatureFinder/__init__.py
@@ -2,5 +2,4 @@
 from importlib.metadata import version
 __version__ = version("PhenoFeatureFinder")
 
-__author__ = 'Marc Galland'
-__email__ = 'm.galland@uva.nl'
+__author__ = "Lissy-Anne Denkers, Marc Galland, Annabel Dekker, Valerio Bianchi and Petra Bleeker"
