### 2024-10-03 5b1aa22
improve introduction text Read the Docs
