### 2024-10-03 02bfe0a
new attempt to generate docstring documentation
