### 2024-10-07 2f16ea6
generate documentation from docstrings
