### 2024-10-07 cc7c563
manually write documentation based on docstrings
