+++ docs/docstrings.rst
@@ -6,19 +6,1561 @@ Documentation
 PhenotypeAnalysis
 #################
 
-.. automodule:: phenofeaturefinder.phenotype_analysis
-    :members:
+A class to analyse data from developmental bioassays and group the samples in distinct phenotypic classes. 
 
+
+Parameters
+----------
+
+bioassay_csv: string
+    A path to a .csv file with the bioassay count data.
+
+
+Attributes
+----------
+
+bioassay_csv
+    The path to the csv file with the input data.
+bioassay
+    A pandas dataframe with the development data.
+sample_id
+    The name of the column that contains the sample identifiers.
+cumulative_data
+    A pandas dataframe with cumulative development data.
+max_counts
+    A pandas dataframe with the maximum number of nymphs developed to or past each developmental stage per group.
+survival_data
+    A pandas dataframe with the number of living nymphs per stage at each time point per group.
+order_of_groups
+    A string with the order in which the groups should be plotted.
+absolute_counts
+    A pandas dataframe with values from max_counts in long format.
+max_relative
+    A pandas dataframe based on absolute_counts, with values relative to values of specified developmental stage.
+
+
+Methods
+-------
+
+reshape_to_wide
+    Reshapes the dataframe from a long to a wide format to make the data accessible for pre-processing. With the counts of each developmental stage in a seperate columns.
+combine_seperately_counted_versions_of_last_recorded_stage
+    Calculates the total number of nymphs developed to the final developmental stage per sample on each timepoint.
+convert_counts_to_cumulative
+    Calculates the total number of nymphs developed to or past each stage on each timepoint.
+correct_cumulative_counts
+    Inner function for convert_counts_to_cumulative().
+create_df_with_max_counts_per_stage
+    Inner function for convert_counts_to_cumulative(). 
+prepare_for_plotting
+    Prepare the order in which the groups should be plotted.
+plot_counts_per_stage
+    Plots the counts per nymphal stage in boxplots.
+plot_development_over_time_in_fitted_model
+    Fits a 3 parameter log-logistic curve to the development over time to a specified stage.
+ll3
+    A three parameter log-logistic function.
+plot_survival_over_time_in_fitted_model
+    Fits a 3 parameter log-normal curve to the number of living nymphs over time.
+hazard
+    A three parameter log-normal function.
+
+
+reshape_to_wide
+***************
+
+Reshapes the dataframe from a long to a wide format to make the data accessible for pre-processing.
+with the counts of each developmental stage in a seperate columns.
+
+
+Usage
+-----
+
+reshape_to_wide(
+    self,
+    sample_id='sample_id',
+    grouping_variable='genotype',
+    developmental_stages='stage',
+    count_values='number',
+    time='day')
+
+
+Parameters
+----------
+
+sample_id: string, default='sample_id'
+    The name of the column that contains the sample identifiers.
+grouping_variable: string, default='genotype'
+    The name of the column that contains the names of the grouping variables.
+    Examples are genotypes or treatments
+developmental_stages: string, default='stage'
+    The name of the column that contains the developmental stages that were scored during the bioassay.
+count_values: string, default='numbers'
+    The name of the column that contains the counts.
+time: string, default='day'
+    The name of the column that contains the time at which bioassay scoring was performed.
+    Examples are the date or the number of days after infection.
+
+Examples
+--------
+
+Example of an input dataframe
+
++-----------+-----------+-------+---------------+-------+
+| sample_id | genotype  | day   | stage         | count |
++===========+===========+=======+===============+=======+
+| mm_1      |   mm      | 5     | eggs          | 45    |
++-----------+-----------+-------+---------------+-------+
+| mm_1      |   mm      | 5     | first_instar  | 0     |
++-----------+-----------+-------+---------------+-------+
[…]
