### 2024-10-07 9d86d61
add version number to documentation
