### 2024-10-07 7bbbcba
Delete .github/workflows/draft-pdf.yml
