### 2024-10-07 3c4954f
add instructions for testing
