### 2024-10-14 #7264 @AdmiralenOla
Dear @petrableeker and @LissyDenkers,

I've reviewed your paper and software. Like the other reviewers, I would like to commend you on your notebook examples. The figures also really helped me understand the intended workflow.

I have filled out the reviewer checklist. Some specific points below:

> **Contribution and authorship:** Has the submitting author (@petrableeker) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Like my co-reviewer @pierrepeterlongo said, the contributions of the submitting author (and of authors 3 and 4) is not clear from the commit history. Author contributions is not included in the paper. Hence, it is not possible to evaluate this point.

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

I checked the box, but found the bullet point confusing. There doesn't appear to be any automated tests, but the functionality of the software can be tested by replicating the examples in the documentation.

> **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
> **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

While this is sufficiently clear on the Github project page and in the code documentation, I would recommend that you clarify the intended use (and user) a bit more in your paper. The paper says that it focuses on insect development but that "different input data with a similar structure could be used". Indeed, for the OmicsAnalysis and FeatureSelection classes it is easy to imagine other uses since it is very similar to GWAS type problems. However, it was hard for me to come up with another use case for the PhenotypeAnalysis class since it specifically deals with the different developmental stages.

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Related to the previous point. It is clearly beyond the scope of the paper to give an overview of all 'omics-phenotype association software. On the other hand, this is probably(?) the only one to specifically deal with developmental stages of insects. The easiest fix here, in my opinion, would be to be more clear about the insect use case. The current title uses "developmental phenotypes" which seems vague.

**Other comments:**

In Figure 1, the arrows make it look like the output from PhenotypeAnalysis and OmicsAnalysis combines into FeatureSelection. Is it not more that the output from PhenotypeAnalysis feeds directly into OmicsAnalysis, in this example Circles and triangles?
