### 2024-10-16 2319d06
rephrase functionality of PhenotypeAnalysis
