+++ README.md
@@ -1,7 +1,6 @@
 [![Documentation Status](https://readthedocs.org/projects/phenofeaturefinder/badge/?version=latest)](https://phenofeaturefinder.readthedocs.io/en/latest/?badge=latest)
 ![PyPI - Version](https://img.shields.io/pypi/v/PhenoFeatureFinder)
 ![PyPI - Python Version](https://img.shields.io/pypi/pyversions/PhenoFeatureFinder)
-![PyPI - Downloads](https://img.shields.io/pypi/dm/PhenoFeatureFinder)
 
 
 
@@ -14,9 +13,9 @@
 * `OmicsAnalysis`
 * `FeatureSelection`
 
-![Overview of the package](documentation/paper/package_figure.png)
+![Overview of the package](user_material/paper/package_figure.png)
 
-`PhenotypeAnalysis` was designed to analyse the development of insects through their larval stages over time in multiple treatments (e.g. hosts). These types of phenotyping analyses can be challenging, due to the many variables involved (e.g. time, developmental stages, replicates, treatments), especially for researchers whose strength does not lie in data analysis. `PhenotypeAnalysis` offers a set of functions to visualise the development while taking into account those different variables, and to perform the necessary data preprocessing steps. From the output, it is easy to manually assign binary phenotypes to your groups (treatments, genotypes, etc), if you want to use it as input for `FeatureSelection`. 
+`PhenotypeAnalysis` was designed to analyse development over time through progressive stages in multiple groups or treatments. This could for example be the development of insects through their larval stages over time in different environments, or disease scores of fungal infections in multiple host plants. These types of phenotyping analyses can be challenging, due to the many variables involved (e.g. time, developmental stages, replicates, treatments), especially for researchers whose strength does not lie in data analysis. `PhenotypeAnalysis` offers a set of functions to visualise the development while taking into account those different variables, and to perform the necessary data preprocessing steps. From the output, it is easy to manually assign binary phenotypes to your groups (treatments, genotypes, etc), if you want to use it as input for `FeatureSelection`. 
 
 With `OmicsAnalysis`, you can filter large untargeted metabolomics datasets and visualise the structure of the data. The filtered data and a corresponding set of binary phenotypes can then be used as input for `FeatureSelection`. With only a few lines of code, the best fitting pipeline to link the phenotypes to metabolic features is created using Automated Machine Learning with `TPOT` and `scikit-learn`.
 
@@ -32,9 +31,9 @@ At this moment, `PhenoFeatureFinder` requires python 3.9.
 
 ## Usage
 
-For each of the classes, you can find a manual with an explanation for all of their functions in the [manuals folder](documentation/manuals/). Alternatively, you can find the documentation of the classes and their functions on [Read the Docs](https://phenofeaturefinder.readthedocs.io).
+For each of the classes, you can find a manual with an explanation for all of their functions in the [manuals folder](user_material/manuals/). Alternatively, you can find the documentation of the classes and their functions on [Read the Docs](https://phenofeaturefinder.readthedocs.io).
 
-If you want to see an example of how `PhenoFeatureFinder` can be used for real-world data, you can take a look at one of the two [examples](documentation/examples/). The first example showcases the use of the [`PhenotypeAnalysis` class](documentation/examples/caddisfly/) for the analysis of the development of caddisfly larvae in four freshwater streams. In the second example, the [`OmicsAnalysis` and `FeatureSelection` classes](documentation/examples/MicroMass/) are used to analyse and select interesting features from a mass spectrometry dataset of a panel of bacterial species.
[…]
