+++ docs/docstrings.rst
@@ -12,14 +12,14 @@ A class to analyse data from developmental bioassays and group the samples in di
 
 
 Parameters
-----------
+**********
 
-bioassay_csv: string
+bioassay_csv: ``string``
     A path to a .csv file with the bioassay count data.
 
 
 Attributes
-----------
+**********
 
 bioassay_csv
     The path to the csv file with the input data.
@@ -34,7 +34,7 @@ max_counts
 survival_data
     A pandas dataframe with the number of living nymphs per stage at each time point per group.
 order_of_groups
-    A string with the order in which the groups should be plotted.
+    A `string` with the order in which the groups should be plotted.
 absolute_counts
     A pandas dataframe with values from max_counts in long format.
 max_relative
@@ -42,7 +42,7 @@ max_relative
 
 
 Methods
--------
+*******
 
 reshape_to_wide
     Reshapes the dataframe from a long to a wide format to make the data accessible for pre-processing. With the counts of each developmental stage in a seperate columns.
@@ -69,14 +69,14 @@ hazard
 
 
 reshape_to_wide
-***************
+---------------
 
 Reshapes the dataframe from a long to a wide format to make the data accessible for pre-processing.
 with the counts of each developmental stage in a seperate columns.
 
 
 Usage
------
+*****
 
 reshape_to_wide(
     self,
@@ -88,23 +88,23 @@ reshape_to_wide(
 
 
 Parameters
-----------
+**********
 
-sample_id: string, default='sample_id'
+sample_id: `string`, default='sample_id'
     The name of the column that contains the sample identifiers.
-grouping_variable: string, default='genotype'
+grouping_variable: `string`, default='genotype'
     The name of the column that contains the names of the grouping variables.
     Examples are genotypes or treatments
-developmental_stages: string, default='stage'
+developmental_stages: `string`, default='stage'
     The name of the column that contains the developmental stages that were scored during the bioassay.
-count_values: string, default='numbers'
+count_values: `string`, default='numbers'
     The name of the column that contains the counts.
-time: string, default='day'
+time: `string`, default='day'
     The name of the column that contains the time at which bioassay scoring was performed.
     Examples are the date or the number of days after infection.
 
 Examples
---------
+********
 
 Example of an input dataframe
 
@@ -133,7 +133,7 @@ Example of a reshaped output dataframe
 
 
 combine_seperately_counted_versions_of_last_recorded_stage
-**********************************************************
+----------------------------------------------------------
 
 Calculates the total number of nymphs developed to the final developmental stage per sample on each timepoint.
 This is used when nymphs in the (late) final nymph stage were removed after each counting moment and/or
@@ -142,7 +142,7 @@ Removal of late last stage nymphs could for example be used to prevent adults fr
 
 
 Usage
------
+*****
 
 combine_seperately_counted_versions_of_last_recorded_stage(
     self,
@@ -157,39 +157,39 @@ combine_seperately_counted_versions_of_last_recorded_stage(
 
 
 Parameters
-----------
+**********
 
-exuviea: string, default='exuviea'
+exuviea: `string`, default='exuviea'
     The name of the column that contains the exuviea counts. 
-late_last_stage: string, default='late_fourth_instar'
+late_last_stage: `string`, default='late_fourth_instar'
     The name of the column that contains the counts of the last developmental stage recorded in the bioassay.
-early_last_stage: string, default='early_fourth_instar'
+early_last_stage: `string`, default='early_fourth_instar'
     The name of the column that contains the counts of the nymphs in early last developmental stage.
     Is used when nymphs counted in late_last_stage were removed after each counting moment during the bioassay.
-new_last_stage: string, default='fourth_instar'
[…]
