+++ docs/index.md
@@ -5,10 +5,8 @@
 :maxdepth: 2
 :hidden:
 
-intro.rst
 docstrings.rst
 changelog.md
 contributing.md
 conduct.md
-autoapi/index
 ```
