+++ docs/index.rst
@@ -12,5 +12,4 @@ Contents
     changelog.md
     contributing.md
     conduct.md
-    autoapi/index
 
