+++ docs/docstrings.rst
@@ -34,7 +34,7 @@ max_counts
 survival_data
     A pandas dataframe with the number of living nymphs per stage at each time point per group.
 order_of_groups
-    A `string` with the order in which the groups should be plotted.
+    A string with the order in which the groups should be plotted.
 absolute_counts
     A pandas dataframe with values from max_counts in long format.
 max_relative
@@ -75,61 +75,61 @@ Reshapes the dataframe from a long to a wide format to make the data accessible 
 with the counts of each developmental stage in a seperate columns.
 
 
-Usage
-*****
+    Usage
+    *****
 
-reshape_to_wide(
-    self,
-    sample_id='sample_id',
-    grouping_variable='genotype',
-    developmental_stages='stage',
-    count_values='number',
-    time='day')
-
-
-Parameters
-**********
-
-sample_id: `string`, default='sample_id'
-    The name of the column that contains the sample identifiers.
-grouping_variable: `string`, default='genotype'
-    The name of the column that contains the names of the grouping variables.
-    Examples are genotypes or treatments
-developmental_stages: `string`, default='stage'
-    The name of the column that contains the developmental stages that were scored during the bioassay.
-count_values: `string`, default='numbers'
-    The name of the column that contains the counts.
-time: `string`, default='day'
-    The name of the column that contains the time at which bioassay scoring was performed.
-    Examples are the date or the number of days after infection.
-
-Examples
-********
-
-Example of an input dataframe
-
-+-----------+-----------+-------+---------------+-------+
-| sample_id | genotype  | day   | stage         | count |
-+===========+===========+=======+===============+=======+
-| mm_1      |   mm      | 5     | eggs          | 45    |
-+-----------+-----------+-------+---------------+-------+
-| mm_1      |   mm      | 5     | first_instar  | 0     |
-+-----------+-----------+-------+---------------+-------+
-| mm_1      |   mm      | 5     | second_instar | 0     |
-+-----------+-----------+-------+---------------+-------+
+    reshape_to_wide(
+        self,
+        sample_id='sample_id',
+        grouping_variable='genotype',
+        developmental_stages='stage',
+        count_values='number',
+        time='day')
 
 
-Example of a reshaped output dataframe
+    Parameters
+    **********
 
-+-----------+-----------+-------+-------+---------------+---------------+---------------+
-| sample_id | genotype  | day   | eggs  | first_instar  | second instar | third_instar  |
-+===========+===========+=======+=======+===============+===============+===============+
-| mm_1      |   mm      | 5     | 45    | 0             | 0             | 0             |
-+-----------+-----------+-------+-------+---------------+---------------+---------------+
-| mm_1      |   mm      | 9     | NA    | 10            | 5             | 0             |
-+-----------+-----------+-------+-------+---------------+---------------+---------------+
-| mm_1      |   mm      | 11    | NA    | 15            | 17            | 4             |
-+-----------+-----------+-------+-------+---------------+---------------+---------------+
+    sample_id: `string`, default='sample_id'
+        The name of the column that contains the sample identifiers.
+    grouping_variable: `string`, default='genotype'
+        The name of the column that contains the names of the grouping variables.
+        Examples are genotypes or treatments
+    developmental_stages: `string`, default='stage'
+        The name of the column that contains the developmental stages that were scored during the bioassay.
+    count_values: `string`, default='numbers'
+        The name of the column that contains the counts.
+    time: `string`, default='day'
+        The name of the column that contains the time at which bioassay scoring was performed.
+        Examples are the date or the number of days after infection.
+
+    Examples
+    ********
+
+    Example of an input dataframe
+
+    +-----------+-----------+-------+---------------+-------+
[…]
