+++ docs/docstrings.rst
@@ -75,8 +75,7 @@ Reshapes the dataframe from a long to a wide format to make the data accessible 
 with the counts of each developmental stage in a seperate columns.
 
 
-    Usage
-    *****
+**Usage**
 
     reshape_to_wide(
         self,
@@ -87,8 +86,7 @@ with the counts of each developmental stage in a seperate columns.
         time='day')
 
 
-    Parameters
-    **********
+**Parameters**
 
     sample_id: `string`, default='sample_id'
         The name of the column that contains the sample identifiers.
@@ -103,33 +101,32 @@ with the counts of each developmental stage in a seperate columns.
         The name of the column that contains the time at which bioassay scoring was performed.
         Examples are the date or the number of days after infection.
 
-    Examples
-    ********
+**Examples**
 
     Example of an input dataframe
 
-    +-----------+-----------+-------+---------------+-------+
-    | sample_id | genotype  | day   | stage         | count |
-    +===========+===========+=======+===============+=======+
-    | mm_1      |   mm      | 5     | eggs          | 45    |
-    +-----------+-----------+-------+---------------+-------+
-    | mm_1      |   mm      | 5     | first_instar  | 0     |
-    +-----------+-----------+-------+---------------+-------+
-    | mm_1      |   mm      | 5     | second_instar | 0     |
-    +-----------+-----------+-------+---------------+-------+
++-----------+-----------+-------+---------------+-------+
+| sample_id | genotype  | day   | stage         | count |
++===========+===========+=======+===============+=======+
+| mm_1      |   mm      | 5     | eggs          | 45    |
++-----------+-----------+-------+---------------+-------+
+| mm_1      |   mm      | 5     | first_instar  | 0     |
++-----------+-----------+-------+---------------+-------+
+| mm_1      |   mm      | 5     | second_instar | 0     |
++-----------+-----------+-------+---------------+-------+
 
 
     Example of a reshaped output dataframe
 
-    +-----------+-----------+-------+-------+---------------+---------------+---------------+
-    | sample_id | genotype  | day   | eggs  | first_instar  | second instar | third_instar  |
-    +===========+===========+=======+=======+===============+===============+===============+
-    | mm_1      |   mm      | 5     | 45    | 0             | 0             | 0             |
-    +-----------+-----------+-------+-------+---------------+---------------+---------------+
-    | mm_1      |   mm      | 9     | NA    | 10            | 5             | 0             |
-    +-----------+-----------+-------+-------+---------------+---------------+---------------+
-    | mm_1      |   mm      | 11    | NA    | 15            | 17            | 4             |
-    +-----------+-----------+-------+-------+---------------+---------------+---------------+
++-----------+-----------+-------+-------+---------------+---------------+---------------+
+| sample_id | genotype  | day   | eggs  | first_instar  | second instar | third_instar  |
++===========+===========+=======+=======+===============+===============+===============+
+| mm_1      |   mm      | 5     | 45    | 0             | 0             | 0             |
++-----------+-----------+-------+-------+---------------+---------------+---------------+
+| mm_1      |   mm      | 9     | NA    | 10            | 5             | 0             |
++-----------+-----------+-------+-------+---------------+---------------+---------------+
+| mm_1      |   mm      | 11    | NA    | 15            | 17            | 4             |
++-----------+-----------+-------+-------+---------------+---------------+---------------+
 
 
 combine_seperately_counted_versions_of_last_recorded_stage
@@ -141,8 +138,7 @@ when exuviea and last instar stage nymphs were counted seperately.
 Removal of late last stage nymphs could for example be used to prevent adults from emerging and escaping.
 
 
-    Usage
-    *****
+**Usage**
 
[…]
