+++ user_material/paper/paper.md
@@ -45,23 +45,29 @@ bibliography: paper.bib
 
 # Summary
 
-Plants interact with their (a)biotic environment through a range of specialised metabolites. They deal with pathogens and pest attack through constitutive or inducible production of specialised metabolites or other defence molecules [@Erb2020; @García-Olmedo1998]. High-throughput “-omics” tools including (untargeted) metabolomics have been successfully implemented in plant biology [@Dalio2021], but the accompanying resistance phenotyping often lacks in robustness [@Song2021]. Resistance is often not a binary trait, but quantitative in nature. To identify features such as defence metabolites involved in a resistance phenotype, we developed a pipeline that includes phenotypic analysis, preprocessing and visualisation of the metabolomics data, and feature prediction through a Machine Learning approach. 
-
-Proliferation of an insect population is affected by various factors, including the chemical composition of the host, and/or the environment [@Ma2022]. In particular, host resistance via hampered larval development is noteworthy, because reducing the speed at which larvae reach the adult stage and produce offspring negatively affects pest-population development [@Maharijaya2019; @Muema2016; @Vengateswari2022]. However, evaluating larval development results in a complex dataset that is challenging to process. Developmental success is based on the number of larvae throughout various larval stages, as well as on the speed of development. 
-
-To identify underlying mechanisms of resistance, the chemical or molecular composition of a plant can be investigated. Proteins and metabolites are commonly analysed through untargeted Mass-Spectrometry, yielding exhaustive profiles generally consisting of many thousands of unannotated features. Often such data displays sparsity, i.e. missing values between datasets, and a low sample-to-feature ratio, adding to the complexity of the analysis [@Kortbeek2021; @Liebal2020]. Tree-based Machine-Learning algorithms (e.g., random forest) are suitable for the analysis of, and feature selection from, untargeted data [@Liebal2020] computing the contribution of each feature in the phenotypic classification. 
-
-`PhenoFeatureFinder` is designed to facilitate the different analyses mentioned above in one pipeline. It can be used for 1) evaluation and visualisation of pest performance over multiple stages and between groups (treatments, genotypes), 2) pre-processing of the omics data, and 3) prediction of features that explain the phenotypic classification. To facilitate usability, each step in the pipeline can also be performed independently, hence has been assigned a class in the package (\autoref{fig1}). Also, although we focus on insect development and the selection of metabolic features causal to the observed phenotype, different input data with a similar structure could be used. `PhenoFeatureFinder` was developed initially for metabolomics data, users can evaluate its fit applying other types of omics data.
+`PhenoFeatureFinder` is designed to facilitate the analyses required to analyse quantitative and/or progressive phenotypic- and omics data, and link those using Machine Learning with the aim to identify causal features, in one package. It can be used for 1) evaluation and visualisation of phenotype progression over multiple stages and between groups (e.g. treatments, genotypes), 2) pre-processing of omics data, and 3) prediction of features that explain the phenotypic classification. To facilitate usability, each step in the pipeline can also be performed independently, hence has been assigned a class in the package (\autoref{fig1}). We provide an example of implementation below that focuses on insect development through time and the selection of metabolic features causal to the observed phenotype, but different input data could be used, provided it has a similar structure. This could be any phenotype that is scored in progressive stages over time. Also, `PhenoFeatureFinder` was developed initially with metabolomics data, but users can evaluate its fit applying other types of omics data.
 
[…]
