### 2024-10-21 #7264 @petrableeker
Dear @AdmiralenOla and @pierrepeterlongo,

Thank you for your feedback. We addressed all issues in the manuscript and the Github repository and we clarify below:

As you both pointed out, the contributions of three of the authors are not visible in the commit history. To clarify the contributions of all authors, we added an ‘Author contributions’ segment to the manuscript and the GitHub repo.

The testing of the software is indeed manual, as described in the GitHub repo and mentioned by @AdmiralenOla. Are these manual testing steps sufficient, or do you have suggestions for how to approach it differently?

As you both mentioned, the target audience, purpose and functionality were not yet clearly described in the manuscript. We rewrote the text based on your feedback and agree that your points indeed clarify our intentions. Also, we clarified this in the GitHub repo as requested by @pierrepeterlongo.

The directories named ‘docs’ and ‘documentation’ were indeed confusing, as remarked by @pierrepeterlongo. The previous ‘documentation’ directory now has the name ‘user-material’.

The readthedocs page no longer has both the index and intro page and now only has the index page. This page indeed contains the same information as the README of the GitHub repo. This decision was made for user friendliness, because with the links, a user can directly find information like the manuals or examples in the GitHub repo without needing prior understanding of the structure of GitHub or Read the Docs.

@pierrepeterlongo thank you for spotting the mistake in the notation of $e_{50}$.

@AdmiralenOla in figure 1, the output of PhenotypeAnalysis and OmicsAnalysis indeed combine into FeatureSelection, as suggested by the arrows, but the confusion is very understandable. We tried to use generic symbols and limit the number of different symbols to prevent confusion, but this was clearly not the best solution. We changed the symbols accordingly.
