### 2024-11-01 #7264 @pierrepeterlongo
Hello @petrableeker

Thank you for your answers and the modifications you made on PhenoFeatureFinder paper, repo, and doc.

The summary and the statement of need are much clearer, and the value of the tool is now obvious from these sections.

The Section "Example implementation" is a bit confusing. I was expecting a "Usage example". We the current presentation, it might be understood as the user has to re-implement PhenoFeatureFinder depending on its usage. Is that the case? Maybe just change the section title to "Use case example"

A new paragraph enables to check the box "state of the field". As mentioned by @AdmiralenOla, this software paper cannot cover the whole state of the art of this vast domain.

The last remaining point is the Contribution and authorship. I did not check the box. Authors contributions are now clearly indicated. However we cannot answer positively to the question "Has the submitting author @petrableeker made major contributions to the software?" However, remember that even though she is the submitting author, Petra Bleeker is the last author of the paper, while the two main contributors, Lissy Denkers and Marc Galland, are the first two authors. So the order of authorship respects the contributions.

Suggestions:

- add hyper-links in page https://github.com/BleekerLab/PhenoFeatureFinder/tree/master/user_material/manuals
