### 2024-11-01 #7264 @jromanowska
@pierrepeterlongo, thank you for checking. It's basically up to the authors to decide whether they deserve the authorship and "contributions" may also be e.g., financial or organizational, not only coding. So you can tick off the box in the checklist :)
