### 2024-11-01 #7264 @jromanowska
@lopeztarifa, have you had time to check the changes the authors made? Do they satisfy the checklist criteria now?
