### 2024-11-04 #7264 @AdmiralenOla
@petrableeker, @LissyDenkers. Thanks for amending with respect to my comments! Sure, if manual tests are good enough for the JOSS requirements they are good enough for me :-)

@jromanowska - I've checked the remaining boxes on my check-list.
