### 2024-11-04 35e5cc1
add hyperlinks to manuals readme
