### 2024-11-04 #7264 @LissyDenkers
@AdmiralenOla, @lopeztarifa and @pierrepeterlongo, thank you for reviewing our work and providing feedback.

@pierrepeterlongo, I’ve implemented your last feedback/suggestions to rename the example section in the manuscript and add hyperlinks to the manuals page.
