### 2024-11-13 #7264 @jromanowska
I've noticed that one reference has wrong capitalized 'r' in the title: "Dose-response analysis using r" (should be capital 'R')
