+++ user_material/paper/paper.bib
@@ -179,7 +179,7 @@
    journal = {PLoS ONE},
    pages = {1-13},
    pmid = {26717316},
-   title = {Dose-response analysis using R},
+   title = {Dose-response analysis using {R}},
    volume = {10},
    year = {2015},
 }
