### 2024-11-15 #7264 @LissyDenkers
@jromanowska, I hope you feel better.

The reference title should now have a capital 'R', thanks for noticing.
I'm working on the "additional author tasks" and will let you know what the archive DOI and version number are when I've got them.
