+++ docs/docstrings.rst
@@ -2,7 +2,7 @@
 Documentation
 *************
 
-Version 0.1.2
+Version 0.1.3
 
 
 PhenotypeAnalysis
