### 2024-11-19 #7264 @LissyDenkers
@jromanowska, all the additional author tasks should be completed:
- The newest version of the manuscript contains updated author affiliations.
- The software release with all changes from the review is version v0.1.3
- This version is archived on Zenodo with DOI 10.5281/zenodo.14179965
