### 2024-11-20 #7264 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
