### 2024-11-20 #7264 @Kevin-Mattheus-Moerman
- [ ] `nessecary` should read `necessary`
- [ ] Check if `timepoint` should read `time point`
- [ ] Check if `high loadings` should read `high loading`
- [ ] The following is not at all a requirement but a recommendation. I would recommend using `\cdot` which produces $\cdot$, rather than `\times` which produces $\times$ (or better still perhaps to leave out the multiplication symbols entirely), since the former is more common in scientific notation and the later looks to many engineers/physicists like a cross product operator. But again, not a requirement, this may be a personal preference.
